FUN-9 Refactor, create product_to_category module
