--- server/src/domain/product/modules/category/category.interface.ts
@@ -1,8 +1,0 @@
-export interface CategoryInterface {
-  id: number;
-  name: string;
-  slug: string;
-}
-
-export interface CategoryToAddInterface
-  extends Omit<CategoryInterface, 'id' | 'slug'> {}
