+++ server/src/domain/product/modules/product/product.controller.ts
@@ -9,7 +9,10 @@ import {
   Query,
 } from '@nestjs/common';
 import { ProductService } from './product.service';
-import { ProductInterface, ProductToAddInterface } from './product.interface';
+import {
+  ProductInterface,
+  ProductToAddInterface,
+} from '../../product.domain.interface';
 
 @Controller('/products')
 export class ProductController {
@@ -20,17 +23,6 @@ export class ProductController {
     @Query() searchParams: Partial<ProductInterface>,
   ): Promise<ProductInterface[]> {
     return await this.productService.getProducts(searchParams);
-  }
-
-  @Get('category/:category_id')
-  async getProductsFromCategory(
-    @Param('category_id') category_id: number,
-    @Query() searchParams: Partial<ProductInterface>,
-  ): Promise<ProductInterface[]> {
-    return await this.productService.getProductsFromCategory(
-      category_id,
-      searchParams,
-    );
   }
 
   @Get('id/:id')
@@ -59,26 +51,5 @@ export class ProductController {
   @Delete(':id')
   async deleteProduct(@Param('id') id: number): Promise<void> {
     await this.productService.deleteProduct(id);
-
-    this.productService.deleteProductFromAllCategories(id);
-  }
-
-  @Post('/categories/:category_id/product/:product_id')
-  async addProductToCategory(
-    @Param('category_id') category_id: number,
-    @Param('product_id') product_id: number,
-  ): Promise<void> {
-    await this.productService.addProductToCategory(category_id, product_id);
-  }
-
-  @Delete('/categories/:category_id/product/:product_id')
-  async deleteProductCategory(
-    @Param('category_id') category_id: number,
-    @Param('product_id') product_id: number,
-  ): Promise<void> {
-    await this.productService.deleteProductFromCategory(
-      category_id,
-      product_id,
-    );
   }
 }
