+++ server/src/domain/product/modules/product/product.module.ts
@@ -2,9 +2,10 @@ import { Module } from '@nestjs/common';
 import { DatabaseModule } from '../../../../modules/database/database.module';
 import { ProductController } from './product.controller';
 import { ProductService } from './product.service';
+import { ProductToCategoryModule } from '../product_to_category/product_to_category.module';
 
 @Module({
-  imports: [DatabaseModule],
+  imports: [DatabaseModule, ProductToCategoryModule],
   controllers: [ProductController],
   providers: [ProductService],
 })
