+++ server/src/domain/product/modules/product_to_category/product_to_category.module.ts
@@ -1,0 +1,12 @@
+import { Module } from '@nestjs/common';
+import { DatabaseModule } from '../../../../modules/database/database.module';
+import { ProductToCategoryController } from './product_to_category.controller';
+import { ProductToCategoryService } from './product_to_category.service';
+
+@Module({
+  imports: [DatabaseModule],
+  exports: [ProductToCategoryService],
+  controllers: [ProductToCategoryController],
+  providers: [ProductToCategoryService],
+})
+export class ProductToCategoryModule {}
