+++ server/src/domain/product/product.domain.registry.ts
@@ -1,0 +1,9 @@
+enum TABLES {
+  P = 'products',
+  C = 'categories',
+  PTC = 'product_to_categories',
+}
+
+export const DB = {
+  TABLES,
+};
